Move default stringifiers to the package level
