+++ client/src/main/java/org/spine3/string/LongStringifier.java
@@ -1,0 +1,64 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package org.spine3.string;
+
+import com.google.common.primitives.Longs;
+
+import java.io.Serializable;
+
+/**
+ * The {@code Stringifier} for the long values.
+ *
+ * @author Illia Shepilov
+ * @author Alexander Yevsyukov
+ */
+final class LongStringifier extends Stringifier<Long> implements Serializable {
+
+    private static final long serialVersionUID = 1;
+
+    private static final LongStringifier INSTANCE = new LongStringifier();
+
+    static LongStringifier instance() {
+        return INSTANCE;
+    }
+
+    @Override
+    protected String toString(Long obj) {
+        return Longs.stringConverter()
+                    .reverse()
+                    .convert(obj);
+    }
+
+    @Override
+    protected Long fromString(String s) {
+        return Longs.stringConverter()
+                    .convert(s);
+    }
+
+    @Override
+    public String toString() {
+        return "Stringifiers.forLong()";
+    }
+
+    private Object readResolve() {
+        return INSTANCE;
+    }
+}
